rename `declaration-property-value-no-deprecated` to `declaration-property-value-keyword-no-deprecated`
