+++ lib/rules/declaration-property-value-keyword-no-deprecated/__tests__/index.mjs
