+++ lib/rules/declaration-property-value-keyword-no-deprecated/index.cjs
@@ -1,11 +1,15 @@
-import { declarationValueIndex } from '../../utils/nodeFieldIndices.mjs';
-import { isPlainObject } from '../../utils/validateTypes.mjs';
-import report from '../../utils/report.mjs';
-import ruleMessages from '../../utils/ruleMessages.mjs';
-import { singleValueColorProperties } from '../../reference/properties.mjs';
-import validateOptions from '../../utils/validateOptions.mjs';
+// NOTICE: This file is generated by Rollup. To modify it,
+// please instead edit the ESM counterpart and rebuild with Rollup (npm run build).
+'use strict';
 
-const ruleName = 'declaration-property-value-no-deprecated';
+const nodeFieldIndices = require('../../utils/nodeFieldIndices.cjs');
+const validateTypes = require('../../utils/validateTypes.cjs');
+const report = require('../../utils/report.cjs');
+const ruleMessages = require('../../utils/ruleMessages.cjs');
+const properties = require('../../reference/properties.cjs');
+const validateOptions = require('../../utils/validateOptions.cjs');
+
+const ruleName = 'declaration-property-value-keyword-no-deprecated';
 
 const messages = ruleMessages(ruleName, {
 	rejected: (value, property) =>
@@ -13,7 +17,7 @@ const messages = ruleMessages(ruleName, {
 });
 
 const meta = {
-	url: 'https://stylelint.io/user-guide/rules/declaration-property-value-no-deprecated',
+	url: 'https://stylelint.io/user-guide/rules/declaration-property-value-keyword-no-deprecated',
 	fixable: true,
 };
 
@@ -42,10 +46,10 @@ const DEPRECATED_COLORS = {
 	threedface: 'ButtonFace',
 	inactivecaptiontext: 'GrayText',
 };
-const COLOR_PROPERTIES = [...singleValueColorProperties.values()].map(
+const COLOR_PROPERTIES = [...properties.singleValueColorProperties.values()].map(
 	(value) => /** @type {[string, Record<String, string>]} */ ([value, DEPRECATED_COLORS]),
 );
-const NAME_TO_VALUE = new Map([
+const PROPERTY_NAME_TO_KEYWORD = new Map([
 	[
 		'appearance',
 		/** @type {string | string[] | Record<string, string>} */ ({
@@ -96,7 +100,7 @@ const rule = (primary) => {
 		if (!validOptions) return;
 
 		root.walkDecls((decl) => {
-			const value = NAME_TO_VALUE.get(decl.prop.toLowerCase());
+			const value = PROPERTY_NAME_TO_KEYWORD.get(decl.prop.toLowerCase());
 
 			if (!value) return;
 
@@ -108,7 +112,7 @@ const rule = (primary) => {
 
 			if (Array.isArray(value) && !value.includes(keyword)) return;
 
-			if (isPlainObject(value)) {
+			if (validateTypes.isPlainObject(value)) {
 				const fixedValue = value[keyword];
 
 				if (!fixedValue) return;
@@ -118,7 +122,7 @@ const rule = (primary) => {
 				};
 			}
 
-			const index = declarationValueIndex(decl);
+			const index = nodeFieldIndices.declarationValueIndex(decl);
 			const endIndex = index + declarationValue.length;
 
 			report({
@@ -138,4 +142,5 @@ const rule = (primary) => {
 rule.ruleName = ruleName;
 rule.messages = messages;
 rule.meta = meta;
-export default rule;
+
+module.exports = rule;
