+++ lib/rules/index.cjs
@@ -125,8 +125,10 @@ const rules = {
 	get 'declaration-property-value-disallowed-list'() {
 		return Promise.resolve().then(() => /*#__PURE__*/_interopNamespaceDefaultOnly(require('./declaration-property-value-disallowed-list/index.cjs'))).then((m) => m.default);
 	},
-	get 'declaration-property-value-no-deprecated'() {
-		return Promise.resolve().then(() => /*#__PURE__*/_interopNamespaceDefaultOnly(require('./declaration-property-value-no-deprecated/index.cjs'))).then((m) => m.default);
+	get 'declaration-property-value-keyword-no-deprecated'() {
+		return Promise.resolve().then(() => /*#__PURE__*/_interopNamespaceDefaultOnly(require('./declaration-property-value-keyword-no-deprecated/index.cjs'))).then(
+			(m) => m.default,
+		);
 	},
 	get 'declaration-property-value-no-unknown'() {
 		return Promise.resolve().then(() => /*#__PURE__*/_interopNamespaceDefaultOnly(require('./declaration-property-value-no-unknown/index.cjs'))).then((m) => m.default);
