+++ lib/rules/index.mjs
@@ -119,8 +119,10 @@ const rules = {
 	get 'declaration-property-value-disallowed-list'() {
 		return import('./declaration-property-value-disallowed-list/index.mjs').then((m) => m.default);
 	},
-	get 'declaration-property-value-no-deprecated'() {
-		return import('./declaration-property-value-no-deprecated/index.mjs').then((m) => m.default);
+	get 'declaration-property-value-keyword-no-deprecated'() {
+		return import('./declaration-property-value-keyword-no-deprecated/index.mjs').then(
+			(m) => m.default,
+		);
 	},
 	get 'declaration-property-value-no-unknown'() {
 		return import('./declaration-property-value-no-unknown/index.mjs').then((m) => m.default);
